refactor scraper modules, reduce files
