+++ scraper/scrape_profs.py
@@ -1,0 +1,3 @@
+import argparse
+from RMPscraper import RMPscraper, save_info
+from RMPmodules import get_ratings
